fix: v2버전으로 파일 업데이트
