--- src/pages/generate/v2/types/loading.ts
@@ -1,9 +1,0 @@
-// 스택 UI
-export interface CarouselItem {
-  rawProductId: number;
-  url: string;
-}
-
-export interface ImageStackResponse {
-  carousels: CarouselItem[];
-}
